New accountclasses

+++ src/main/java/com/example/banksimulation/CurrentAccount.java
@@ -1,7 +1,7 @@
 package com.example.banksimulation;
 
 public class CurrentAccount extends Account{
-    CurrentAccount(int accountNumber, int accountBalance, Customer accountHolder) {
+    CurrentAccount(int accountNumber, double accountBalance, Customer accountHolder) {
         super(accountNumber, accountBalance, accountHolder);
     }
